Составлено 9 тестов для покрытия приложения BooksCollector
